requirements.txt
###### Requirements without Version Specifiers ######
selenium
python-dotenv